requirements.txt
attrs==20.1.0
certifi==2020.6.20
colorcet==2.0.2
cycler==0.10.0
filelock==3.0.12
importlib-metadata==1.7.0
iniconfig==1.0.1
kiwisolver==1.1.0
matplotlib==3.3.1
more-itertools==8.5.0
numpy==1.18.5
packaging==20.4
pandas==1.1.1
param==1.9.3
pathlib2==2.3.5
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pyct==0.4.6
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
scipy==1.4.1
six==1.15.0
tabulate==0.8.7
toml==0.10.1
